pyproject.toml
[project]
name = "event-pusher"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.29.0",
    "google-api-python-client>=2.128.0",
]

[dependency-groups]
dev = [
    "boto3>=1.34.0",
    "google-auth-oauthlib>=1.2.0",
    "pytest>=8.2.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
